requirements.txt
diffusers
torch
gdown
torchvision
pipeline
transformers
accelerate
safetensors
sentencepiece
spaces
peft
gradio
pillow